add storybook to the project, with a story for the EpisodeCard component
Storybook is a way of being able to render and test out your components
in isolation from an app, to see how they respond to changing the props
passed into it. You can then publish these components to other people,
and the Storybook site that you generate can give designers and
implementers direction for how the component should be used.